move repository for example task under its domain model directory
