+++ library-modules/domain-models/task-example/repositories/mappers/task-mapper.ts
@@ -1,5 +1,5 @@
 import { DbTask } from "/library-modules/apis/example-tasks/models/DbTask";
-import { Task } from "/library-modules/domain-models/Task";
+import { Task } from "../../Task";
 
 export function mapDbTasksToTasks(tasks: DbTask[]): Task[] {
   return tasks.map((task) => {
